Remove empty subsections (#22)
This PR addresses an issue where certain resources, such as arb-tn and arb-tq, would appear on BIEL with subcontents that had no links.  For example:

<img src="https://user-images.githubusercontent.com/1095575/150559390-c4db42a5-7ba5-4f72-ab06-8aa6cd6d2c98.png" width="250px" />

The issue is not that resources are missing, but rather that after filtering out unavailable resources, the empty subcontents were hanging around.  (In this case, we removed links to the Scripture Builder, which doesn't yet work with non-Scripture resources.)

This PR is comprised of the following changes:

- Add `filterEmptySubcontents()` to functions.js, which finds empty subcontents sections and removes them.
- Update the `massage()` function in index.js to call `filterEmptySubcontents()`.
- Remove unused debug `console.log()` statements.
- Remove individual batch files and replace them with a single makefile.